+++ packages/expo-maps/CHANGELOG.md
@@ -6,7 +6,11 @@
 
 ### 🎉 New features
 
+- [android] Add `latitudeDelta` and `longitudeDelta` to `onCameraMove` event payload. ([#45327](https://github.com/expo/expo/pull/45327) by [@nishan](https://github.com/intergalacticspacehighway))
+
 ### 🐛 Bug fixes
+
+- [android] Trigger `onCameraMove` on initial mount when map is ready. ([#45327](https://github.com/expo/expo/pull/45327) by [@nishan](https://github.com/intergalacticspacehighway))
 
 ### 💡 Others